+++ src/PSRule/Resources/ViewStrings.Designer.cs
@@ -160,6 +160,15 @@ namespace PSRule.Resources {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Tag.
+        /// </summary>
+        internal static string Tag {
+            get {
+                return ResourceManager.GetString("Tag", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to TargetName.
         /// </summary>
         internal static string TargetName {